requirements.txt
# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
matplotlib==3.7.1
mrmr-selection==0.2.6
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
scipy==1.10.1
PyYAML==6.0
ReliefF==0.1.2
skrebate==0.62
matplotlib-venn==0.11.9
mdfs==1.5.2
isort==5.12.0
black==23.3.0
flake8==6.0.0
setuptools~=67.6.1